requirements.txt
fastapi 
uvicorn 
pydantic 
sqlalchemy 
asyncpg 
aio-pika 
redis 
httpx 
python-dotenv 
structlog